This code sucks!
The recursion of the function is not good enough to understand
that if the root is smaller than child then swap.

This way of the code is not enough and needs to be changed.

Unncessary call to buildheap() function is not needed.
Nor the mapheapify call to function every time.